GH-45 extract raspberry specific to seperate module
This task it the first in the sequence that provides the flashing
feature for RK1 compute modules. In this commit raspberry specific code
is moved to a seperate file. Secondly the runway is implemented to
support multiple "drivers" that enable read/write to internal storage of
compute modules. e.g. Writing new firmware to eMMC storage of a module

see:
https://github.com/turing-machines/BMC-Firmware/issues/45