Implemented RK1 fwupdate support.
This work fixes #45:
    https://github.com/turing-machines/BMC-Firmware/issues/45

Implemented discovery of an RK1 module as well as Implemented the
firmware update read/write trait. Note that its implemented in a
blocking way. #12 is opened to track any work to improve on this.

This implementation expects the RK1 to be put into maskedROM mode.
Writing to eMMC starts from sector 0.