rockusb_fwudate: rk3588 spl_loader
Load an spl that inits ram and setup the usb flashing environment when
The device is booted into MaskRom mode. Without this step, writing a
image to the device will fail.

This commit makes use of rockusb branch of svenrademakers. Hopefully
the patch associated with this CL will be submitted upstream soon.