bmcd: Implement graceful shutdown
This commit enables a graceful shutdown of the bmcd application.
(typically gets triggered on a kill signal). The event_listener
referenced the `BmcApplication,` preventing it from dropping. The fix
entails a cancellation_token that will shut down this listening task.

Fixes #107 partly,
https://github.com/turing-machines/BMC-Firmware/issues/107 ,th  bmcd
application does shutdown, which drops the database connection. However,
it does not guarantee that the filesystem buffers are flushed to flash.
Secondly, a graceful reboot command is still required.

Lastly, the behavior of writing/reading to USB transports changed
instead of ignoring erroneous reads/writes. The functions now return
directly when an error occurs.