calculate_checksum: fix bug
This issue occurs when the written image size is not equal to a product
of the internal read buffer size. Improved the code so that the exact
amount of bytes are read when calculating a checksum. This means the
that the last read will be trimmed to match the file size. Instead of
reading a full buffer.