bmcd: IntoLegacyResponse
Implemented boilerplate code to transform `anyhow::Result` and
`HttpResponses` into `LegacyResponses`. The goal of the `LegacyResponse`
type is to enforce a specific layout of the response body. Because
legacy responders must specify a `LegacyResponse` before it gets
transformed into a `HttpResponse`.

Changed API port from 8080 to 80. The former is a more industry comform
port. However to not break legacy tools we are stuck with port 80 port
for the `legacy API`

Refactored out the Arc wrapper around `BmcApplication`. moved the
event_listener to a seperate module. This should make the relation
between the `EventListener` and `BmcApplication` more visible.