bmcd: listen on IPv6 & increase KeepAlive
Prevent the TCP connection from closing during a flashing process. When
the receive buffer is full, byte chunk requests will be blocked until
new space becomes available. This caused connection resets, resulting in
an abort of the flashing process. This happens, for example, during the
configuration phase of the node which typically takes longer as 5 sec.
default keep-alive for HTTP1.1 was 5 sec. This CL align sets the
keep-alive timeout to the value of the underlaying OS.