rockusb_fwudate: Update bootmode detection
After discussion with the maintainers, decided is to keep the Bootmode
detection logic local to this project. There is no generic solution
possible that works for all rockusb devices. Detection is based on
behavior of the loader that is installed on the rockchip device, no
spec exists that guarantees a certain behavior of this loader.
see https://github.com/collabora/rockchiprs/pull/22

Updated the rockusb dependency to the main branch, as one of the latest
patches are required.