bmcd: expose API "info" endpoint
This is a poor man's way to introduce a versioning in the 'BMC API'.
Version info can be retrieved by calling "/bmc/api/info". This endpoint
is exposed both on the HTTP and HTTPS port.

Secondly this commit contains some small updates on the response
handling in the legacy module.