streaming_data_service: fix get error msg deadlock
When an error occurs during a file transfer, method
`return_transfer_error` always falls back on the default error message
'transfer canceled'.

The issues is fixed by not holding the streaming_data_service state lock
when waiting for the error message to bubble up. This gives the
streaming_data_service the opportunity to update the error state.