serial_service: Prep work for byte stream support
This commit aims to refactor the serial_service component. The
functionality stays the same, therefore no changes in the external API as well.

An additional hook is added in the serial handler that will be used in
subsequent commits to implement a byte stream object with.

Implemented the `open_channel()` function which was marked as `todo!()`.
This channel is a asynchronous bi-directional bytes transport. Purposed
to handle the communication to the various Nodes on the TuringPi board.