serial over websocket
This commit contains a websocket implementation for a given node. This
websocket implementation communicates with `bytes` messages. A watchdog
is implemented to monitor client pings.

a websocket can be opened using the following API call:
`/api/bmc/serial/ws?node=0`, which opens a serial connection over a
websocket to node 1.