app_persistency: small async refactor
The 'sync_all()' call on Drop of the AppPersistency camouflages the
error messages in case of a critical error. This makes debugging more
hard. By making the 'sync_all()' call synchronous, the problem is
solved. This has minimal to no drawbacks since the filesystem calls inside the
function do not benefit rusts async scheduling.