app_persistency: fix regression
There was a blocking `.sync_all()` on drop to guarantee all data was
synced to the disk. Scheduling async tasks on a blocking context in
tokio is somewhat of a bad practice.

As a solution, the task is scheduled conventionally with 'tokio::spawn'.
This means there is no guarantee the task is executed before the exit of
the program. This is mitigated with a fast writeback timeout on changes.