workaround: supply default node info
This commit fixes the symptom of a serialization issue, allegedly after
an upgrade from 2.0.5.
My hypothesis is that the problem is introduced due to the upgrade of
the rust compiler. As this could change the way the node-info structure is serialized.

Please note that the storage files of are not forward compatible. So
switching back from newer versions can also introduce problems.