config: fix issues upgrading user config
When a user edit's the bmcd config file, a copy gets overlayed over the original. This means that newly added config keys in the original config are not visible anymore. By migrating to rusts 'config' crate, code is added that merges correctly default keys with user overriden keys from the overlay file.